+++ samples/expo/babel.config.js
@@ -1,6 +1,0 @@
-module.exports = function (api) {
-  api.cache(false);
-  return {
-    presets: ['babel-preset-expo'],
-  };
-};